requirements.txt
livekit-agents>=0.8.7
livekit-plugins-openai>=0.8.1
livekit-plugins-silero>=0.6.4
python-dotenv~=1.0